LayoutTests/fast/forms/color/color-setrangetext.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../js/resources/js-test-pre.js"></script>
<script src="../resources/common-setrangetext.js"></script>
</head>
<body>
<script>
description("Test setRangeText() method is not available in color inputs.");
runTestsShouldFail("input", { type: "color" });
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
