LayoutTests/fast/forms/color/color-setrangetext-expected.txt

Test setRangeText() method is not available in color inputs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Running tests on input with attributes: {"type":"color"}

element.value = '0123456789XYZ'
PASS element.setRangeText('ABC', 0, 0) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS element.value is "#000000"
PASS successfullyParsed is true

TEST COMPLETE
